+++ config/index.js
@@ -19,6 +19,11 @@ const defaultRoot = resolve(process.env.HOME, ".cynomys", "protal");
 const [num, dataDir, logDir] = [cpus().length, resolve(defaultRoot, "data"), resolve(defaultRoot, "logs")];
 
 siteConfig.set("version", "1.0.8"); // 设置服务器版本
+siteConfig.set("ssl", { // 安全设置
+  enable: true, // 是否开启ssl设置
+  name: "cynomy", // ssl公私钥文件名
+// 存放位置,  path: "~/.ssh"  // 不写则默认位置
+});
 siteConfig.set("remote", { // 远程节点设置
   enable : false, // 是否开启远程
   host: "namo", // 远程服务器ip